Relocate system identification tools-in-progress to solvers directory and rename the files to conventional under_scores style rather than FunkyCaps.
