+++ drake/solvers/CMakeLists.txt
@@ -31,7 +31,7 @@ if(MATLAB_FOUND)
 endif()
 
 set(optimization_files)
-list(APPEND optimization_files MathematicalProgram.cpp MobyLCP.cpp)
+list(APPEND optimization_files MathematicalProgram.cpp MobyLCP.cpp system_identification.cpp)
 pods_find_pkg_config(nlopt)
 if (nlopt_FOUND)
   list(APPEND optimization_files NloptSolver.cpp)
@@ -46,7 +46,7 @@ else()
 endif(snopt_c_FOUND)
 add_library_with_exports(LIB_NAME drakeOptimization SOURCE_FILES ${optimization_files})
 set_target_properties(drakeOptimization PROPERTIES COMPILE_FLAGS -fPIC)
-drake_install_headers(MathematicalProgram.h Optimization.h)
+drake_install_headers(MathematicalProgram.h Optimization.h system_identification.h)
 pods_install_libraries(drakeOptimization)
 if (nlopt_FOUND)
   pods_use_pkg_config_packages(drakeOptimization nlopt)
